refactor: code deduplication
Abstract the avatar fetching, caching, and error handling into a single function.
